Split sensors into separate modules
