+++ src/sensors/power_consumption.rs
@@ -1,0 +1,44 @@
+//! Module providing power consumption sensor functionality.
+
+use super::{LocalisedSensorMetadata, SensorMetadata, SensorTemplate, Sensors};
+use serde::{Deserialize, Serialize};
+use std::convert::TryInto;
+
+#[derive(Serialize, Deserialize, Default, Debug, Clone, PartialEq)]
+pub struct PowerConsumptionSensor {
+    #[serde(flatten)]
+    pub metadata: LocalisedSensorMetadata,
+    pub unit: String,
+    pub value: f64,
+}
+
+#[derive(Debug, Clone)]
+pub struct PowerConsumptionSensorTemplate {
+    pub metadata: SensorMetadata,
+    pub unit: String,
+}
+
+impl TryInto<PowerConsumptionSensor> for PowerConsumptionSensorTemplate {
+    type Error = Box<dyn std::error::Error>;
+
+    fn try_into(self) -> Result<PowerConsumptionSensor, Self::Error> {
+        Ok(PowerConsumptionSensor {
+            metadata: self.metadata.try_into()?,
+            unit: self.unit,
+            ..PowerConsumptionSensor::default()
+        })
+    }
+}
+
+impl SensorTemplate for PowerConsumptionSensorTemplate {
+    fn try_to_sensor(
+        &self,
+        value_str: &str,
+        sensors: &mut Sensors,
+    ) -> Result<(), Box<dyn std::error::Error>> {
+        let mut sensor: PowerConsumptionSensor = self.clone().try_into()?;
+        sensor.value = value_str.parse::<f64>()?;
+        sensors.power_consumption.push(sensor);
+        Ok(())
+    }
+}
